scrapper/requirements.txt
playwright
aiohttp
groq
openai
python-dotenv
